+++ Source/Scene/GlobeSurfaceShaderSet.js
@@ -181,17 +181,9 @@ define([
 
             if (enableClippingPlanes) {
                 fs.defines.push('ENABLE_CLIPPING_PLANES');
-                fs.defines.push('CLIPPING_PLANES_TEXTURE_WIDTH ' + ClippingPlaneCollection.TEXTURE_WIDTH);
-
                 if (ClippingPlaneCollection.useFloatTexture(frameState.context)) {
-                    fs.defines.push('CLIPPING_PLANES_TEXTURE_HEIGHT_FLOAT ' + ClippingPlaneCollection.TEXTURE_HEIGHT_FLOAT);
                     fs.defines.push('FLOAT_CLIPPING_PLANES');
                 }
-
-                /*
-                if (unionClippingRegions) {
-                    fs.defines.push('UNION_CLIPPING_REGIONS');
-                }*/
             }
 
             var computeDayColor = '\