+++ Source/Scene/Model.js
@@ -2850,13 +2850,13 @@ define([
         };
     }
 
-    function createClippingPlanesLengthRangeUnionFunction(model) {
+    function createClippingPlanesRangeFunction(model) {
         return function() {
             var clippingPlanes = model.clippingPlanes;
             if (!defined(clippingPlanes) || !clippingPlanes.enabled) {
-                return Cartesian4.ZERO;
+                return Cartesian2.ZERO;
             }
-            return clippingPlanes.lengthRangeUnion;
+            return clippingPlanes.range;
         };
     }
 
@@ -2986,11 +2986,16 @@ define([
             uniformMap = combine(uniformMap, {
                 gltf_color : createColorFunction(model),
                 gltf_colorBlend : createColorBlendFunction(model),
-                gltf_clippingPlanesLengthRangeUnion: createClippingPlanesLengthRangeUnionFunction(model),
                 gltf_clippingPlanes: createClippingPlanesFunction(model),
                 gltf_clippingPlanesEdgeStyle: createClippingPlanesEdgeStyleFunction(model),
                 gltf_clippingPlanesMatrix: createClippingPlanesMatrixFunction(model)
             });
+
+            if (!ClippingPlaneCollection.useFloatTexture(context)) {
+                uniformMap = combine(uniformMap, {
+                    gltf_clippingPlanesRange: createClippingPlanesRangeFunction(model)
+                });
+            }
 
             // Allow callback to modify the uniformMap
             if (defined(model._uniformMapLoaded)) {
@@ -3791,26 +3796,30 @@ define([
     }
 
     function modifyShaderForClippingPlanes(shader, clippingPlaneCollection, context) {
+        var usingRgbaTexture = !ClippingPlaneCollection.useFloatTexture(context);
+
         shader = ShaderSource.replaceMain(shader, 'gltf_clip_main');
+        shader += Model._getClippingFunction(clippingPlaneCollection) + '\n';
+        if (usingRgbaTexture) {
             shader +=
-            Model._getClippingFunction(clippingPlaneCollection) +
-            '\n' +
-                'uniform vec4 gltf_clippingPlanesLengthRangeUnion; \n' +
-                'uniform sampler2D gltf_clippingPlanes; \n' +
-                'uniform vec4 gltf_clippingPlanesEdgeStyle; \n' +
-                'uniform mat4 gltf_clippingPlanesMatrix; \n' +
-                'void main() \n' +
-                '{ \n' +
-                '    gltf_clip_main(); \n' +
-            '    float clipDistance = clip(gl_FragCoord, gltf_clippingPlanes, gltf_clippingPlanesMatrix' + (ClippingPlaneCollection.useFloatTexture(context) ? ');\n' : ', gltf_clippingPlanesLengthRangeUnion.yz);\n') +
-            '    vec4 clippingPlanesEdgeColor = vec4(1.0); \n' +
-            '    clippingPlanesEdgeColor.rgb = gltf_clippingPlanesEdgeStyle.rgb; \n' +
-            '    float clippingPlanesEdgeWidth = gltf_clippingPlanesEdgeStyle.a; \n' +
-            '    if (clipDistance > 0.0 && clipDistance < clippingPlanesEdgeWidth) \n' +
-                '    { \n' +
-            '        gl_FragColor = clippingPlanesEdgeColor;\n' +
-                '    } \n' +
-                '} \n';
+            'uniform vec2 gltf_clippingPlanesRange; \n';
+        }
+        shader +=
+            'uniform sampler2D gltf_clippingPlanes; \n' +
+            'uniform vec4 gltf_clippingPlanesEdgeStyle; \n' +
+            'uniform mat4 gltf_clippingPlanesMatrix; \n' +
+            'void main() \n' +
+            '{ \n' +
+            '    gltf_clip_main(); \n' +
+        '    float clipDistance = clip(gl_FragCoord, gltf_clippingPlanes, gltf_clippingPlanesMatrix' + (usingRgbaTexture ? ', gltf_clippingPlanesRange);\n' : ');\n') +
+        '    vec4 clippingPlanesEdgeColor = vec4(1.0); \n' +
+        '    clippingPlanesEdgeColor.rgb = gltf_clippingPlanesEdgeStyle.rgb; \n' +
+        '    float clippingPlanesEdgeWidth = gltf_clippingPlanesEdgeStyle.a; \n' +
+        '    if (clipDistance > 0.0 && clipDistance < clippingPlanesEdgeWidth) \n' +
+            '    { \n' +
+        '        gl_FragColor = clippingPlanesEdgeColor;\n' +
+            '    } \n' +
+            '} \n';
         return shader;
     }